Refactoring of Large capital profile unit tests
Some of the files where becoming unmanageable, therefore, I've broken
these out into smaller files.
