+++ test/unit/apps/companies/controllers/investments/large-capital-profile/details-edit.test.js
@@ -1,17 +1,17 @@
 const investorTypeTransformed = require('~/test/unit/data/companies/investments/metadata/investor-type-transformed.json')
-const noCompanyProfile = require('~/test/unit/data/companies/investments/large-capital-profile-empty.json')
-const investorType = require('~/test/unit/data/companies/investments/metadata/investor-type.json')
 const requiredChecksConducted = require('~/test/unit/data/companies/investments/metadata/required-checks-conducted.json')
+const investorType = require('~/test/unit/data/companies/investments/metadata/investor-type.json')
 const companyProfile = require('~/test/unit/data/companies/investments/large-capital-profile.json')
-const companyMock = require('~/test/unit/data/companies/minimal-company.json')
 const advisersMock = require('~/test/unit/data/advisers/advisers.json')
+
 const buildMiddlewareParameters = require('~/test/unit/helpers/middleware-parameters-builder.js')
-const { cloneDeep, pullAll } = require('lodash')
+const companyMock = require('~/test/unit/data/companies/minimal-company.json')
+const { cloneDeep } = require('lodash')
 const config = require('~/config')
 
 const controller = require('~/src/apps/companies/apps/investments/large-capital-profile/controllers')
 
-describe('Company Investments - large capital profile', () => {
+describe('Company Investments - Large capital profile - Investor details', () => {
   describe('renderProfile', () => {
     const commonTests = (profile) => {
       it('should call the render function once', () => {
@@ -28,78 +28,7 @@ describe('Company Investments - large capital profile', () => {
       })
     }
 
-    context('when the company does not have a company profile', () => {
-      beforeEach(async () => {
-        nock(config.apiRoot)
-          .get(`/v4/large-investor-profile?investor_company_id=${companyMock.id}`)
-          .reply(200, noCompanyProfile)
-
-        this.middlewareParameters = buildMiddlewareParameters({
-          company: companyMock,
-        })
-
-        await controller.renderProfile(
-          this.middlewareParameters.reqMock,
-          this.middlewareParameters.resMock,
-          this.middlewareParameters.nextSpy,
-        )
-      })
-
-      commonTests(undefined)
-    })
-
-    context('when the company has a new unedited profile', () => {
-      beforeEach(async () => {
-        nock(config.apiRoot)
-          .get(`/v4/large-investor-profile?investor_company_id=${companyMock.id}`)
-          .reply(200, companyProfile)
-
-        this.middlewareParameters = buildMiddlewareParameters({
-          company: companyMock,
-        })
-
-        await controller.renderProfile(
-          this.middlewareParameters.reqMock,
-          this.middlewareParameters.resMock,
-          this.middlewareParameters.nextSpy,
-        )
-      })
-
-      commonTests({
-        editing: undefined,
-        id: companyProfile.results[0].id,
-        investorDetails: {
-          incompleteFields: 5,
-          investorType: {
-            text: null,
-            value: null,
-          },
-          globalAssetsUnderManagement: {
-            value: null,
-          },
-          investableCapital: {
-            value: null,
-          },
-          investorDescription: {
-            value: '',
-          },
-          requiredChecks: {
-            adviser: null,
-            date: null,
-            type: null,
-            value: null,
-          },
-        },
-        investorRequirements: {
-          incompleteFields: 9,
-        },
-        location: {
-          incompleteFields: 3,
-        },
-      })
-    })
-
-    context('when the company has a profile and the user is editing the "Investor details" section', () => {
+    context('when the user is editing the "Investor details" section', () => {
       beforeEach(async () => {
         const clonedCompanyProfile = cloneDeep(companyProfile)
         const profile = clonedCompanyProfile.results[0]
@@ -247,95 +176,9 @@ describe('Company Investments - large capital profile', () => {
         },
         investorRequirements: {
           incompleteFields: 9,
-        },
-        location: {
-          incompleteFields: 3,
-        },
-      })
-    })
-
-    context('when the user has previously saved all fields within "Investor details"', () => {
-      beforeEach(async () => {
-        const clonedCompanyProfile = cloneDeep(companyProfile)
-        const profile = clonedCompanyProfile.results[0]
-
-        pullAll(profile.incomplete_details_fields, [
-          'investor_type',
-          'global_assets_under_management',
-          'investable_capital',
-          'investor_description',
-          'required_checks_conducted',
-        ])
-
-        profile.investor_type = {
-          id: '80168d31-fa91-494e-9ad5-b9255e01b5da',
-          name: 'Asset manager',
-        }
-        profile.global_assets_under_management = 1000
-        profile.investable_capital = 2000
-        profile.investor_description = 'Lorem ipsum dolor sit amet.'
-        profile.required_checks_conducted = {
-          name: 'Cleared',
-          id: '02d6fc9b-fbb9-4621-b247-d86f2487898e',
-        }
-        profile.required_checks_conducted_by = {
-          name: 'Holly Collins',
-          id: '379f390a-e083-4a2c-9cea-e3b9a08606a7',
-        }
-        profile.required_checks_conducted_on = '2019-04-29'
-
-        nock(config.apiRoot)
-          .get(`/v4/large-investor-profile?investor_company_id=${companyMock.id}`)
-          .reply(200, clonedCompanyProfile)
-
-        this.middlewareParameters = buildMiddlewareParameters({
-          company: companyMock,
-        })
-
-        await controller.renderProfile(
-          this.middlewareParameters.reqMock,
-          this.middlewareParameters.resMock,
-          this.middlewareParameters.nextSpy,
-        )
-      })
-
-      commonTests({
-        editing: undefined,
-        id: 'a9e7afa4-1079-422a-b24e-a77f3ba80375',
-        investorDetails: {
-          incompleteFields: 0,
-          investorType: {
-            text: 'Asset manager',
-            value: '80168d31-fa91-494e-9ad5-b9255e01b5da',
+          dealTicketSizes: {
+            value: [],
           },
-          globalAssetsUnderManagement: {
-            value: 1000,
-          },
-          investableCapital: {
-            value: 2000,
-          },
-          investorDescription: {
-            value: 'Lorem ipsum dolor sit amet.',
-          },
-          requiredChecks: {
-            type: {
-              id: '02d6fc9b-fbb9-4621-b247-d86f2487898e',
-              name: 'Cleared',
-            },
-            date: '2019-04-29',
-            adviser: {
-              id: '379f390a-e083-4a2c-9cea-e3b9a08606a7',
-              name: 'Holly Collins',
-            },
-            value: [
-              'Cleared',
-              'Date of most recent background checks: 29 04 2019',
-              'Person responsible for most recent background checks: Holly Collins',
-            ],
-          },
-        },
-        investorRequirements: {
-          incompleteFields: 9,
         },
         location: {
           incompleteFields: 3,
