+++ test/unit/apps/companies/controllers/investments/large-capital-profile/transformers/investor-details-to-api.test.js
